[refactor] use `this.element` directly
current implementation uses `document.getElementById` with `this.elementId` for targeting input element. this causes an issue with using angle-bracket invocation of `phone-input` in conjunction with passed `id`. Ember components allow us to avoid the `getElementById` call altogether and just use the in-scope `this.element`, which resolves the issue with angle-bracket implementation, and otherwise leaves functionality unchanged.

- updates element targeting to use `this.element`