Disable automatic deployment of the website
When deploying the website for the main branch, the asset path of the
chunks dynamically imported (for example
`import('intl-tel-input/build/js/intlTelInput.js')`) is wrong.

For instance, webpack tries to load
https://qonto.github.io/ADDON_DOCS_ROOT_URL/assets/chunk.819.cee670bcd021ff0a6cf6.js
instead of
https://qonto.github.io/ember-phone-input/versions/master/assets/chunk.819.cee670bcd021ff0a6cf6.js

Why? There is a ember-cli-deploy plugin in ember-cli-addon-docs. Among
other things, this plugin replaces the placeholder
`ADDON_DOCS_ROOT_URL` in the generated assets files by
`/ember-phone-input/versions/master/`.
For some reason, the plugin doesn't replace the placeholder for
dynamically imported chunks.