Use `:root` prefix to ensure correct `background-image` paths (#506)
Using embroider does not appear to use the same CSS import order as a regular Ember CLI build, and we haven't found a way to enforce it. While using `:root` to increase the specificity is certainly not great, it is probably better than using `!important` and appears to be the best workaround for now until we can find a better solution.